feat: donate page, finish about us and cleanups
