+++ package.json
@@ -32,6 +32,7 @@
     "react-dom": "^17.0.1",
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.31.1",
+    "react-rotating-text": "^1.4.1",
     "tailwindcss": "^3.0.23",
     "yup": "^0.32.11"
   },
