+++ src/components/index.js
@@ -7,3 +7,4 @@ export * from "./br";
 export * from "./modal";
 export * from "./newsletter";
 export * from "./partners";
+export * from "./roadmaps";
