+++ src/components/partners/partners.jsx
@@ -5,8 +5,8 @@ import React from "react";
 
 const Partners = () => {
   return (
-    <Container>
-      <div className=" px-8 md:py-14 py-10 rounded-lg md:mt-[122px] mt-[59px] border border-blue bg-white bg-opacity-5 ">
+    <Container className="text-center">
+      <div className=" px-8 md:py-14 py-10 rounded-lg  border border-blue bg-white bg-opacity-5 ">
         {partners.map(partner => {
           return <Partner key={partner.name} {...partner} />;
         })}
