+++ src/templates/home/index.js
@@ -1,0 +1,2 @@
+export * from "./header";
+export * from "./featured-stories";
